started to add some styling to profile views, added login filters to a few functions in both Page and Profile controller

+++ FinalProject/app/views/Page/pagePage.php
@@ -9,14 +9,15 @@
         <div id="page">
             <div id="navBar">
                 <?php 
-                    if ($data['star'] == false) {
-                        echo "<a href=\"" . BASE . "/Star/add/" . $data['page']->page_id . "\"><i class=\"far fa-star\"></i></a>";
-                    } else {
-                        echo "<a href=\"" . BASE . "/Star/delete/" . $data['page']->page_id . "\"><i class=\"fas fa-star\"></i></a>";
+                    if ($data['star'] != null) {
+                        if ($data['star'] == false) {
+                            echo "<a href=\"" . BASE . "/Star/add/" . $data['page']->page_id . "\"><i class=\"far fa-star\"></i></a>";
+                        } else {
+                            echo "<a href=\"" . BASE . "/Star/delete/" . $data['page']->page_id . "\"><i class=\"fas fa-star\"></i></a>";
+                        }
                     }
                 ?>
                 <a href="<?= BASE ?>/Page/index/<?= $_SESSION['profile_id']?>"> Return to page list</a>
-                <a href="<?= BASE ?>/Comment/add/<?= $data['page']->page_id?>"> Write comment</a> <br /> <br />
                 <a href="<?=BASE?>/Default/logout">Logout</a>
             </div>
 
@@ -28,11 +29,10 @@
             </div>
 
             <div id="comments">
-                <br />
-                <label>Comment section: </label> <br />
+                <h4>Comment section:</h4>
                 <!--list of comments for this page -->
-                <br />
-
+                <a href="<?= BASE ?>/Comment/add/<?= $data['page']->page_id?>" id="addComment"> Add comment</a> <br /> <br />
+                    
                 <?php 
                     foreach ($data['comments'] as $comment) {
                         echo "$comment->comment_text <a href=\"" . BASE . "/Comment/delete/" . $comment->comment_id . "\">delete</a> <br />";
